third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-then-register.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="testutils.js"></script>
<body>
<script>
if (fork()) {
    // The controlling parent frame
    description('Tests destroying a context, then trying to register a new custom element.');
    jsTestIsAsync = true;
    successfullyParsed = true;
} else {
    // The child frame

    var proto = Object.create(HTMLElement.prototype);
    destroyContext();

    try {
        document.registerElement('x-a', {prototype: proto});
        log('FAIL expected register to throw an exception');
    } catch (ex) {
        log('PASS caught expected exception "' + ex + '"');
    }
    done();
}
</script>
