src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/registration-context-delete-then-register-expected.txt

Tests destroying a context, then trying to register a new custom element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS destroyed context
PASS caught expected exception "TypeError: document.registerElement is not a function"
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
